requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
gliner>=0.2.0
faker>=22.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
